fix: File refactoring bugs
Group all created file into Authentication directory
