+++ frontend/lib/Authentication/ForgotPassword.dart
@@ -1,16 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:mapsnap_fe/SignIn.dart';
-import 'package:mapsnap_fe/SignUp.dart';
-import 'package:mapsnap_fe/VerifyEmail.dart';
-import 'Widget/passwordForm.dart';
-import 'Widget/normalForm.dart';
-import 'Widget/outline_IconButton.dart';
+import 'SignUp.dart';
+import 'VerifyEmail.dart';
 import 'Finish.dart';
+import 'package:mapsnap_fe/Widget/passwordForm.dart';
+import 'package:mapsnap_fe/Widget/normalForm.dart';
+import 'package:mapsnap_fe/Widget/outline_IconButton.dart';
 
 
 
-class CreateNewPassword extends StatelessWidget {
-  const CreateNewPassword({super.key});
+class ForgotPassword extends StatelessWidget {
+  const ForgotPassword({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -34,9 +33,9 @@ class CreateNewPassword extends StatelessWidget {
                   child: TextButton.icon(
                     label: Text(""),
                     icon: Image.asset(
-                      "assets/Login/BackButton.png",
-                      width: 50,
-                      height: 50,
+                        "assets/Login/BackButton.png",
+                        width: 50,
+                        height: 50,
                     ),
                     onPressed: () {
                       Navigator.pop(context);
@@ -47,33 +46,31 @@ class CreateNewPassword extends StatelessWidget {
               Padding(
                 padding: const EdgeInsets.only(top: 200),
                 child: Text(
-                  "Create New Password",
+                  "Forgot Password",
                   style: TextStyle(
-                      fontSize: 30,
-                      fontWeight: FontWeight.bold
+                    fontSize: 30,
+                    fontWeight: FontWeight.bold
                   ),
                 ),
               ),
               Padding(
                 padding: const EdgeInsets.only(top: 20),
                 child: Text(
-                  "Please enter and confirm your new password."
-                      "You will need to login after you reset.",
+                  "No worries! Enter your email address below and we "
+                      "will send you a code to reset password.",
                   textAlign: TextAlign.center,
                   style: TextStyle(
-                    fontSize: 15,
+                      fontSize: 15,
                   ),
                 ),
               ),
               // Các ô nhập liệu
               Padding(
                 padding: const EdgeInsets.only(top: 30),
-                child: passwordForm(label:'Password'),
+                child: normalForm(label:'Email'),
               ),
-              Padding(
-                padding: const EdgeInsets.only(top: 30),
-                child: passwordForm(label:'Confirm Password'),
-              ),
+              // passwordForm(label:'Password'),
+              // SizedBox(height: 20),
               Expanded(child: Container()),
 
               // Nút Register
@@ -81,9 +78,9 @@ class CreateNewPassword extends StatelessWidget {
                 padding: const EdgeInsets.only(bottom: 40),
                 child: ElevatedButton(
                   onPressed: () {
-                    Navigator.pushReplacement(
+                    Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => SignIn()),
+                      MaterialPageRoute(builder: (context) => VerifyEmail()),
                     );
                   },
                   style: ElevatedButton.styleFrom(
