+++ frontend/lib/Authentication/SignIn.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:mapsnap_fe/ForgotPassword.dart';
-import 'package:mapsnap_fe/SignUp.dart';
-import 'Widget/passwordForm.dart';
-import 'Widget/normalForm.dart';
-import 'Widget/outline_IconButton.dart';
+import 'ForgotPassword.dart';
+import 'SignUp.dart';
 import 'Finish.dart';
+import 'package:mapsnap_fe/Widget/passwordForm.dart';
+import 'package:mapsnap_fe/Widget/normalForm.dart';
+import 'package:mapsnap_fe/Widget/outline_IconButton.dart';
