+++ frontend/lib/Authentication/SignUp.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:mapsnap_fe/SignIn.dart';
-import 'Widget/passwordForm.dart';
-import 'Widget/normalForm.dart';
-import 'Widget/outline_IconButton.dart';
+import 'SignIn.dart';
+import 'package:mapsnap_fe/Widget/passwordForm.dart';
+import 'package:mapsnap_fe/Widget/normalForm.dart';
+import 'package:mapsnap_fe/Widget/outline_IconButton.dart';
 import 'Finish.dart';
